Darrers ajustos càmera

+++ 2DGame/02-Bubble/02-Bubble/main.cpp
@@ -38,7 +38,7 @@ int main(void)
 		return -1;
 
 	/* Create a windowed mode window and its OpenGL context */
-	window = glfwCreateWindow(384*3.5, 208*3.5, "PANG_VJ", NULL, NULL);
+	window = glfwCreateWindow(int(384*3.5), int(208*3.5), "PANG_VJ", NULL, NULL);
 	if (!window)
 	{
 		glfwTerminate();
